Updated a lot (see notes)
Changed aux daq code to save data to a construct and write that as binary to the SD card. This file must be unpacked with BinFileTranslator.py before being operated on by our preexisting programs. NOTE BinFileTranslator.py is a quick and dirty program made by chatgpt. It should be updated and incorporated into the data translator tool somehow.