Finished Space Concious Data Collection Upgrade
I got the code from earlier working and tested with a generated file.
Average polling rate was 62.9 with expected polling rate of about 64.
This discrepancy could very likely be due to the ADS or some other,
non-code related issue, but we should monitor this for the future.
p-Variance was incredibly low at 0.09