MASSIVE CHANGES
Changed aux daq file output to binary. The aux daq now creates a folder
with the time stamp that contains information on the binary output. The
DAQ.py tool reads this file to decode the binary, and throws away all
duplicate data points. This translation process is unfortunately rather
slow, but we'll hopefully find a way to speed it up in the future.